Added a cast. Should not be necessary but some old GCCs like me to do this.
git-svn-id: svn://svn.code.sf.net/p/peano/code/trunk/src@2512 aa977e5b-e928-4c84-aa66-c90e40304b12

+++ peano/utils/UserInterface.cpp
@@ -179,7 +179,7 @@ int peano::utils::UserInterface::getMemoryUsageMB() {
 void peano::utils::UserInterface::writeHeader(const std::string& experimentName) {
   _log.info("writeHeader()", experimentName );
   _log.info("writeHeader()", _PeanoHeader );
-  _log.info("writeHeader()", "revision: " + std::to_string(PEANO_VERSION) );
+  _log.info("writeHeader()", "revision: " + std::to_string(static_cast<int>(PEANO_VERSION)) );
   #ifdef Parallel
   int numberOfProcesses = tarch::parallel::Node::getInstance().getNumberOfNodes();
   #else
